Add a Makefile
Time to go old school. When using Docker, we can't simply use tox or
manage.py command as commands should run inside the container and not on
the host. To resolve this, we recommend using 'docker-compose run --rm
web tox', which will run tox in the container. Having to remember these
rather esoteric commands is annoying though (particularly when I haven't
run it recently and therefore can't find the command with '<Ctrl>+R'). A
series of make targets make this a heck of a lot easier for us.

Note that this isn't ideal and won't entirely remove the need to run the
above command. This is mostly due to how make is designed: namely, we
need to pass positional arguments but each positional argument in make
should be a target. This does, however, cover the biggest use cases and
is therefore worth doing.

Signed-off-by: Stephen Finucane <stephen@that.guru>