README: Use pyup.io badge instead of requires.io
requires.io has an open issue whereby it doesn't consider Python version
in assessing the latest version of a package available. pyup.io also has
the same issue but we've already worked around this by ignoring the
offending lines. Seeing as we're already using that elsewhere, double
down on it.

Signed-off-by: Stephen Finucane <stephen@that.guru>