tox: Disable W504 ("line break after binary operator")
This was introduced in a recent version of 'pycodestyle'. The
documentation notes [1] that it is mutually exclusive with W503, which
we do enforce, suggesting that we disable one or the other. Avoid the
churn and stick to the older rule, which I personally prefer.

[1] http://pycodestyle.pycqa.org/en/latest/intro.html#error-codes

NOTE(stephenfin): This isn't technically required here, but it does help
us maintain a passing 'pep8' tox target on this branch which is
generally useful for catching things like SyntaxErrors.

Signed-off-by: Stephen Finucane <stephen@that.guru>
(cherry picked from commit 498a8eeaa6c9d1dca8ab29a845325cff1463e815)