travis: Stop testing PostgreSQL 11
Travis seems to be doing something really weird with PG11 since it was
released a few weeks ago. This is currently breaking our CI and can't
continue. It doesn't seem possible to mark this as an expected failure
so simply remove it. We can re-add this once Travis gains proper support
for this version via their addons.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Suggested-by: Daniel Black <daniel@linux.ibm.com>
Acked-by: Daniel Axtens <dja@axtens.net>