REST: Handle JSON requests
This was raising an attribute error when switching tests to use JSON
bodies instead of form-data.

  AttributeError: 'dict' object has no attribute '_mutable'

The easy fix is to check if it's a dictionary and avoid the mutability
check if so.

NOTE(daxtens): This is needed for JSON bodies sent by regular users, not
just the tests.

Signed-off-by: Stephen Finucane <stephen@that.guru>
(cherry picked from commit dc48fbce99efe7d13987a3f510f7dee389636eba)