travis, tox: only do coverage testing in py27
Currently Travis calls `tox -e coverage` unconditionally. However,
the environment has py27 basepython, so all the runs only generate
py27 coverage!

Rather than try to untangle that, just run the coverage when run
in a py27 travis environment. This makes things faster for no
loss of coverage. It means that codecov has nothing to submit for
the py3x environments, but that's no real loss: it would otherwise
submit lots of duplicate data.

We could try to improve coverage by running coverage for 27 and 3x,
but given that 27 is going away, don't stress at this point.

Signed-off-by: Daniel Axtens <dja@axtens.net>