docs: set the GID in the .env file
The .env file needs the UID and GID for a working docker-compose build.

If you follow the current instructions, the docker build fails with a
clear error message: 'You must define GID in .env'

It is still good to update documentation to reduce the burden on new
contributors to run into this build failure first.

Signed-off-by: Pranav Annam <pranavannam@gmail.com>
[simply provide bash command in docs, reworked commit message]
Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>
Signed-off-by: Daniel Axtens <dja@axtens.net>